+++ nvim/.config/nvim/lua/config/options.lua
@@ -3,6 +3,12 @@
 -- Add any additional options here
 
 vim.opt.scrolloff = 10
+
+-- Allow project-local .nvim.lua / .nvimrc for per-repo overrides
+-- with safe mode to restrict potentially dangerous commands
+vim.opt.exrc = true
+vim.opt.secure = true
+
 vim.o.guicursor = "n-v-c:block,i:blinkon0"
 
 -- Disable autoformat globally; opt-in per filetype